Create Catapult
One of the tasks is to throw six velcro covered balls at a mammoth. 
First thing we thought about was a catapult, made up by a servo and something spoon/cup-like.
Because it can't throw all the balls at the same place, we need to make the robot move after a ball
has been thrown, so while the catapult returns to its start position it will move (using a line tracker) so that it aims at a free part of the mammoth, where it will load the next ball and shoot.
We have now thought that there is enough space on the mammoth to do it parallel like this, but we might have to change the height it hits the mammoth by changing the angles the Catapult throws the balls at.